pyproject.toml
[project]
name = "run-run-rest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "authlib>=1.6.10",
    "fastapi>=0.135.3",
    "google-cloud-firestore>=2.26.0",
    "google-genai>=1.73.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "pydantic>=2.13.0",
    "python-dotenv>=1.2.2",
    "python-json-logger>=4.1.0",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
